refactored and changed some styles
