+++ src/store/CalendarContext.tsx
@@ -1,6 +1,24 @@
 import * as React from 'react'
 import { useState } from 'react'
 
+const examples: string[] = [
+  'AE Business Meeting',
+  'Holiday',
+  'Quarter End',
+  'Casual Day',
+  'Pool party',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest',
+  'hot dog eating contest'
+]
+
 const CalendarContext = React.createContext<CalendarStoreValue | undefined>(
   undefined
 )
@@ -13,6 +31,11 @@ interface CalendarStoreValue {
   dayClickCount: number
   selectedDate: undefined | Date
   toggleBarOnDateClick: (num: number, date?: any) => void
+  selected: string[]
+  events: string[]
+  handleChange: (event: { target: { value: any } }) => void
+  handleNone: () => void
+  handleAll: () => void
 }
 
 export const useCalendarContext = () => {
@@ -28,6 +51,26 @@ const CalendarStore = ({ children }: any) => {
   const [yearView, setYearView] = useState(false)
   const [dayClickCount, setDayClickCount] = useState(0)
   const [selectedDate, setSelectedDate] = useState<undefined | Date>(undefined)
+  const [selected, setSelected] = React.useState<string[]>([])
+
+  const handleChange = (event: { target: { value: any } }) => {
+    const value = event.target.value
+    const s: string = value
+    const list = [...selected]
+    const index = list.indexOf(s)
+    index === -1 ? list.push(value) : list.splice(index, 1)
+    setSelected(list)
+  }
+
+  const handleAll = () => {
+    setSelected(examples)
+    return
+  }
+
+  const handleNone = () => {
+    setSelected([])
+    return
+  }
 
   const changeView = (date?: Date) => {
     setYearView(!yearView)
@@ -48,7 +91,12 @@ const CalendarStore = ({ children }: any) => {
     changeView: changeView,
     dayClickCount: dayClickCount,
     selectedDate: selectedDate,
-    toggleBarOnDateClick: toggleBarOnDateClick
+    toggleBarOnDateClick: toggleBarOnDateClick,
+    selected: selected,
+    events: examples,
+    handleChange: handleChange,
+    handleNone: handleNone,
+    handleAll: handleAll
   }
 
   return (
